+++ amethyst_animation/src/resources.rs
@@ -13,6 +13,7 @@ use std::time::Duration;
 /// Blend method for sampler blending
 #[derive(Clone, Copy, Debug, PartialOrd, PartialEq, Eq, Hash)]
 pub enum BlendMethod {
+    /// Simple linear blending
     Linear,
 }
 
@@ -135,6 +136,7 @@ where
 /// Only required for animations which target more than a single node or entity.
 #[derive(Debug, Clone)]
 pub struct AnimationHierarchy<T> {
+    /// A mapping between indices and entities
     pub nodes: FnvHashMap<usize, Entity>,
     m: marker::PhantomData<T>,
 }
@@ -373,6 +375,7 @@ pub struct SamplerControlSet<T>
 where
     T: AnimationSampling,
 {
+    /// The samplers in this set.
     pub samplers: Vec<SamplerControl<T>>,
 }
 
@@ -624,6 +627,7 @@ impl<T> AnimationControl<T>
 where
     T: AnimationSampling,
 {
+    /// Creates a new `AnimationControl`
     pub fn new(
         animation: Handle<Animation<T>>,
         end: EndControl,
@@ -685,6 +689,7 @@ pub struct AnimationControlSet<I, T>
 where
     T: AnimationSampling,
 {
+    /// The animation set.
     pub animations: Vec<(I, AnimationControl<T>)>,
     pub(crate) deferred_animations: Vec<DeferredStart<I, T>>,
 }
@@ -876,6 +881,7 @@ where
     I: Eq + Hash,
     T: AnimationSampling,
 {
+    /// The mapping between `I` and the animation handles.
     pub animations: FnvHashMap<I, Handle<Animation<T>>>,
 }