+++ amethyst_renderer/src/pass/pbm/interleaved.rs
@@ -8,6 +8,7 @@ use cam::{ActiveCamera, Camera};
 use error::Result;
 use gfx::pso::buffer::ElemStride;
 use gfx_core::state::{Blend, ColorMask};
+use hidden::{Hidden, HiddenPropagate};
 use light::Light;
 use mesh::{Mesh, MeshHandle};
 use mtl::{Material, MaterialDefaults};
@@ -70,6 +71,8 @@ where
         Read<'a, AssetStorage<Texture>>,
         ReadExpect<'a, MaterialDefaults>,
         Option<Read<'a, Visibility>>,
+        ReadStorage<'a, Hidden>,
+        ReadStorage<'a, HiddenPropagate>,
         ReadStorage<'a, MeshHandle>,
         ReadStorage<'a, Material>,
         ReadStorage<'a, GlobalTransform>,
@@ -107,6 +110,8 @@ where
             tex_storage,
             material_defaults,
             visibility,
+            hidden,
+            hidden_prop,
             mesh,
             material,
             global,
@@ -119,7 +124,9 @@ where
 
         match visibility {
             None => {
-                for (mesh, material, global) in (&mesh, &material, &global).join() {
+                for (mesh, material, global, _, _) in
+                    (&mesh, &material, &global, !&hidden, !&hidden_prop).join()
+                {
                     draw_mesh(
                         encoder,
                         effect,