+++ amethyst_renderer/src/pass/pbm/separate.rs
@@ -8,6 +8,7 @@ use cam::{ActiveCamera, Camera};
 use error::Result;
 use gfx::pso::buffer::ElemStride;
 use gfx_core::state::{Blend, ColorMask};
+use hidden::{Hidden, HiddenPropagate};
 use light::Light;
 use mesh::{Mesh, MeshHandle};
 use mtl::{Material, MaterialDefaults};
@@ -73,6 +74,8 @@ impl<'a> PassData<'a> for DrawPbmSeparate {
         Read<'a, AssetStorage<Texture>>,
         ReadExpect<'a, MaterialDefaults>,
         Option<Read<'a, Visibility>>,
+        ReadStorage<'a, Hidden>,
+        ReadStorage<'a, HiddenPropagate>,
         ReadStorage<'a, MeshHandle>,
         ReadStorage<'a, Material>,
         ReadStorage<'a, GlobalTransform>,
@@ -132,6 +135,8 @@ impl Pass for DrawPbmSeparate {
             tex_storage,
             material_defaults,
             visibility,
+            hidden,
+            hidden_prop,
             mesh,
             material,
             global,
@@ -145,8 +150,15 @@ impl Pass for DrawPbmSeparate {
 
         match visibility {
             None => {
-                for (joint, mesh, material, global) in
-                    (joints.maybe(), &mesh, &material, &global).join()
+                for (joint, mesh, material, global, _, _) in (
+                    joints.maybe(),
+                    &mesh,
+                    &material,
+                    &global,
+                    !&hidden,
+                    !&hidden_prop,
+                )
+                    .join()
                 {
                     draw_mesh(
                         encoder,